+++ src/main/java/com/martak/adventofcode/intcodecomputer/model/argmode/RelativeArgMode.java
@@ -2,15 +2,13 @@ package com.martak.adventofcode.intcodecomputer.model.argmode;
 
 import com.martak.adventofcode.intcodecomputer.model.Context;
 
-import java.math.BigInteger;
-
 public class RelativeArgMode implements ArgMode {
 
     @Override
-    public BigInteger getValue(Context context, int pointer) {
-        BigInteger[] code = context.getCode();
+    public long getValue(Context context, int pointer) {
+        long[] code = context.getCode();
         int relativeBase = context.getRelativeBase();
-        int address = relativeBase + code[pointer].intValue();
-        return code[address];
+        int address = relativeBase + (int) code[pointer];
+        return (int) code[address];
     }
 }
